Add appservice package and move ApplicationService into it.
